requirements.txt
tensorflow
streamlit
scikit-learn
python-dotenv
kafka-python 
pandas
evidently==0.6.4